Move 'create_pre_udf_table' function to warehouse module
